Split automaton native methods into separate classes. Made agg automation able to run against multiple open windows.
