+++ GuiAutomation/NativeMethods.cs
@@ -87,277 +87,4 @@ namespace MatterHackers.GuiAutomation
 
         public abstract void Type(string textToType);
     }
-
-    public class AggInputMethods : NativeMethods
-    {
-        Point2D currentMousePosition;
-		AutomationRunner automationRunner;
-		public bool DrawSimulatedMouse = true;
-
-		public override ImageBuffer GetCurrentScreen()
-        {
-            throw new NotImplementedException();
-        }
-
-		public  AggInputMethods(AutomationRunner automationRunner, bool drawSimulatedMouse)
-		{
-			this.DrawSimulatedMouse = drawSimulatedMouse;
-			this.automationRunner = automationRunner;
-		}
-
-		public override Point2D CurrentMousPosition()
-        {
-            SystemWindow.AllOpenSystemWindows[0].Invalidate();
-            return currentMousePosition;
-        }
-
-        SystemWindow currentlyHookedWindow = null;
-
-		public override void Dispose()
-		{
-			if (currentlyHookedWindow != null)
-			{
-				currentlyHookedWindow.DrawAfter -= DrawMouse;
-			}
-		}
-
-		public override void SetCursorPosition(int x, int y)
-        {
-            SystemWindow systemWindow = SystemWindow.AllOpenSystemWindows[SystemWindow.AllOpenSystemWindows.Count - 1];
-            if(currentlyHookedWindow != systemWindow)
-            {
-                if (currentlyHookedWindow != null)
-                {
-                    currentlyHookedWindow.DrawAfter -= DrawMouse;
-                }
-                currentlyHookedWindow = systemWindow;
-                if (currentlyHookedWindow != null && DrawSimulatedMouse)
-                {
-                    currentlyHookedWindow.DrawAfter += DrawMouse;
-                }
-            }
-            currentMousePosition = new Point2D(x, y);
-            Point2D windowPosition = AutomationRunner.ScreenToSystemWindow(currentMousePosition, systemWindow);
-            if (LeftButtonDown)
-            {
-                MouseEventArgs aggEvent = new MouseEventArgs(MouseButtons.Left, 1, windowPosition.x, windowPosition.y, 0);
-                UiThread.RunOnIdle(() => systemWindow.OnMouseMove(aggEvent));
-            }
-            else
-            {
-                MouseEventArgs aggEvent = new MouseEventArgs(MouseButtons.None, 1, windowPosition.x, windowPosition.y, 0);
-                UiThread.RunOnIdle(() => systemWindow.OnMouseMove(aggEvent));
-            }
-        }
-
-        private void DrawMouse(GuiWidget drawingWidget, DrawEventArgs e)
-        {
-			automationRunner.RenderMouse(currentlyHookedWindow, e.graphics2D);
-        }
-
-        public override void CreateMouseEvent(int dwFlags, int dx, int dy, int cButtons, int dwExtraInfo)
-        {
-            // figure out where this is on our agg windows
-            // for now only send mouse events to the top most window
-            //foreach (SystemWindow systemWindow in SystemWindow.OpenWindows)
-            SystemWindow systemWindow = SystemWindow.AllOpenSystemWindows[SystemWindow.AllOpenSystemWindows.Count - 1];
-            {
-                Point2D windowPosition = AutomationRunner.ScreenToSystemWindow(currentMousePosition, systemWindow);
-                if(systemWindow.LocalBounds.Contains(windowPosition))
-                {
-                    MouseButtons mouseButtons = MapButtons(cButtons);
-                    // create the agg event
-                    if (dwFlags == MOUSEEVENTF_LEFTDOWN)
-                    {
-                        MouseEventArgs aggEvent = new MouseEventArgs(mouseButtons, 1, windowPosition.x, windowPosition.y, 0);
-                        // send it to the window
-                        if (LeftButtonDown)
-                        {
-                            UiThread.RunOnIdle(() => systemWindow.OnMouseMove(aggEvent));
-                        }
-                        else
-                        {
-                            UiThread.RunOnIdle(() => systemWindow.OnMouseDown(aggEvent));
-                        }
-                    }
-                    else if(dwFlags == MOUSEEVENTF_LEFTUP)
-                    {
-                        MouseEventArgs aggEvent = new MouseEventArgs(mouseButtons, 0, windowPosition.x, windowPosition.y, 0);
-                        // send it to the window
-                        UiThread.RunOnIdle(() => systemWindow.OnMouseUp(aggEvent));
-                    }
-                    else if(dwFlags == MOUSEEVENTF_RIGHTDOWN)
-                    {
-
-                    }
-                    else if (dwFlags == MOUSEEVENTF_RIGHTUP)
-                    {
-
-                    }
-                    else if (dwFlags == MOUSEEVENTF_MIDDLEDOWN)
-                    {
-
-                    }
-                    else if (dwFlags == MOUSEEVENTF_MIDDLEUP)
-                    {
-
-                    }
-                }
-            }
-
-            base.CreateMouseEvent(dwFlags, dx, dy, cButtons, dwExtraInfo);
-        }
-
-        private MouseButtons MapButtons(int cButtons)
-        {
-            switch (cButtons)
-            {
-                case MOUSEEVENTF_LEFTDOWN:
-                case MOUSEEVENTF_LEFTUP:
-                    return MouseButtons.Left;
-
-                case MOUSEEVENTF_RIGHTDOWN:
-                case MOUSEEVENTF_RIGHTUP:
-                    return MouseButtons.Left;
-
-                case MOUSEEVENTF_MIDDLEDOWN:
-                case MOUSEEVENTF_MIDDLEUP:
-                    return MouseButtons.Left;
-            }
-
-            return MouseButtons.Left;
-        }
-
-        public override void Type(string textToType)
-        {
-            SystemWindow systemWindow = SystemWindow.AllOpenSystemWindows[SystemWindow.AllOpenSystemWindows.Count - 1];
-
-            foreach (char character in textToType)
-            {
-                //UiThread.RunOnIdle(() => systemWindow.OnKeyDown(aggKeyEvent));
-                //Keyboard.SetKeyDownState(aggKeyEvent.KeyCode, true);
-
-                KeyPressEventArgs aggKeyPressEvent = new KeyPressEventArgs(character);
-                UiThread.RunOnIdle(() => systemWindow.OnKeyPress(aggKeyPressEvent));
-
-                //widgetToSendTo.OnKeyUp(aggKeyEvent);
-                //Keyboard.SetKeyDownState(aggKeyEvent.KeyCode, false);
-            }
-        }
-    }
-
-#if !__ANDROID__
-    public class WindowsInputMethods : NativeMethods
-    {
-		// P/Invoke declarations
-        [DllImport("gdi32.dll")]
-        static extern bool BitBlt(IntPtr hdcDest, int xDest, int yDest, int wDest, int hDest, IntPtr hdcSource, int xSrc, int ySrc, CopyPixelOperation rop);
-        [DllImport("user32.dll")]
-        static extern bool ReleaseDC(IntPtr hWnd, IntPtr hDc);
-        [DllImport("gdi32.dll")]
-        static extern IntPtr DeleteDC(IntPtr hDc);
-        [DllImport("gdi32.dll")]
-        static extern IntPtr DeleteObject(IntPtr hDc);
-        [DllImport("gdi32.dll")]
-        static extern IntPtr CreateCompatibleBitmap(IntPtr hdc, int nWidth, int nHeight);
-        [DllImport("gdi32.dll")]
-        static extern IntPtr CreateCompatibleDC(IntPtr hdc);
-        [DllImport("gdi32.dll")]
-        static extern IntPtr SelectObject(IntPtr hdc, IntPtr bmp);
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetDesktopWindow();
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetWindowDC(IntPtr ptr);
-
-		[DllImport("User32.Dll")]
-		public static extern long SetCursorPos(int x, int y); 
-
-		[DllImport("user32.dll")]
-		public static extern void mouse_event(int dwFlags, int dx, int dy, int cButtons, int dwExtraInfo);
-
-        public override void CreateMouseEvent(int dwFlags, int dx, int dy, int cButtons, int dwExtraInfo)
-        {
-            mouse_event(dwFlags, dx, dy, cButtons, dwExtraInfo);
-        }
-
-        public override void SetCursorPosition(int x, int y)
-        {
-            SetCursorPos(x, y);
-        }
-
-        public override Point2D CurrentMousPosition()
-        {
-            Point2D mousePos = new Point2D(System.Windows.Forms.Control.MousePosition.X, System.Windows.Forms.Control.MousePosition.Y);
-            return mousePos;
-        }
-
-		public override void Dispose()
-		{
-		}
-
-		public override ImageBuffer GetCurrentScreen()
-		{
-			ImageBuffer screenCapture = new ImageBuffer();
-
-			Size sz = System.Windows.Forms.Screen.PrimaryScreen.Bounds.Size;
-			IntPtr hDesk = GetDesktopWindow();
-			IntPtr hSrce = GetWindowDC(hDesk);
-			IntPtr hDest = CreateCompatibleDC(hSrce);
-			IntPtr hBmp = CreateCompatibleBitmap(hSrce, sz.Width, sz.Height);
-			IntPtr hOldBmp = SelectObject(hDest, hBmp);
-			bool b = BitBlt(hDest, 0, 0, sz.Width, sz.Height, hSrce, 0, 0, CopyPixelOperation.SourceCopy | CopyPixelOperation.CaptureBlt);
-			Bitmap bmpScreenCapture = Bitmap.FromHbitmap(hBmp);
-			SelectObject(hDest, hOldBmp);
-			DeleteObject(hBmp);
-			DeleteDC(hDest);
-			ReleaseDC(hDesk, hSrce);
-
-			//bmpScreenCapture.Save("bitmapsave.png");
-
-			screenCapture = new ImageBuffer(bmpScreenCapture.Width, bmpScreenCapture.Height, 32, new BlenderBGRA());
-			BitmapData bitmapData = bmpScreenCapture.LockBits(new Rectangle(0, 0, bmpScreenCapture.Width, bmpScreenCapture.Height), System.Drawing.Imaging.ImageLockMode.ReadWrite, bmpScreenCapture.PixelFormat);
-
-			int offset;
-			byte[] buffer = screenCapture.GetBuffer(out offset);
-			int bitmapDataStride = bitmapData.Stride;
-			int backBufferStrideInBytes = screenCapture.StrideInBytes();
-			int backBufferHeight = screenCapture.Height;
-			int backBufferHeightMinusOne = backBufferHeight - 1;
-
-			unsafe
-			{
-				byte* bitmapDataScan0 = (byte*)bitmapData.Scan0;
-				fixed (byte* pSourceFixed = &buffer[offset])
-				{
-					byte* pSource = bitmapDataScan0 + bitmapDataStride * backBufferHeightMinusOne;
-					byte* pDestBuffer = pSourceFixed;
-					for (int y = 0; y < screenCapture.Height; y++)
-					{
-						int* pSourceInt = (int*)pSource;
-						pSourceInt -= (bitmapDataStride * y / 4);
-
-						int* pDestBufferInt = (int*)pDestBuffer;
-						pDestBufferInt += (backBufferStrideInBytes * y / 4);
-
-						for (int x = 0; x < screenCapture.Width; x++)
-						{
-							pDestBufferInt[x] = pSourceInt[x];
-						}
-					}
-				}
-			}
-
-			bmpScreenCapture.UnlockBits(bitmapData);
-
-			bmpScreenCapture.Dispose();
-
-			return screenCapture;
-		}
-
-        public override void Type(string textToType)
-        {
-            System.Windows.Forms.SendKeys.SendWait(textToType);
-        }
-    }
-#endif
 }
